Don't set `bed3_to_vcf_index` unless it exists
This code currently returns a temp file pointer for `bed3_to_vcf_index`
even in cases where the file's empty and therefore invalid. This causes
problems because, as the following code snippet shows, Python will throw
an exception when it tries to read an empty tempfile.

    import tempfile
    fpath = tempfile.mktemp()
    with open(fpath, 'rb') as f:
        pass

On the other hand, as one might expect, the following code runs successfully.

    import tempfile
    fpath = tempfile.mktemp()

    with open(fpath, 'w') as f:
        f.write('foo\n')

    with open(fpath, 'rb') as f:
        pass